Move IProjectXmlAccessor/MSBuildXmlAccess to managed
Needed to push other components down into the managed layer.
